requirements (1).txt
gradio 
openai==0.28
google-cloud-texttospeech
python